Python_Project/requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
chromadb
faiss-cpu
langchain_google_genai
langchain-community
pdf2image
youtube_transcript_api
pathlib
langchain_openai
langdetect
nltk
langchain_groq
groq
google-generativeai
phidata 
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
groq
openai
sqlalchemy
pgvector
psycopg[binary]
pypdf
streamlit
youtube_tools